refactor(writer): unify generic arguments and inheritance clause emission
